Remove why-rust topic from A module
The "A-foundations" module included the "Why rust?" topic, which is part
of the "0-intro" module.

The "Rust intro" track include those modules, so the track is including
the topic twice, making the track with repetitive content.

Moreover, the "Why rust?" topic was not sync between the two modules and
the one that was up to date is the one in the "0-intro" module.

This commit address the issue, deleting the topic from the
"A-foundations" module because it make more sense to belong to the
"0-intro" module, besides it was not up to date because was referencing
an issue which is already closed.