modmod: Don't add empty slice if slides.md is empty
Modmod tool was adding an empty slide if `slides.md` was empty. This is
the case for the topic `mods/0-intro/topics/setup`.

This commit check if the slide contents is empty after trimming it and
doesn't write anything. Before it was writing `---` which was
introducing an empty slide in the final Slidev Markdown file.